Simplify: extract control visibility timers (#16 slice 4)
Move the player overlay control visibility state, auto-hide timeout, PiP hide/re-arm
tracking, and timeout cleanup into useControlVisibilityTimers. The overlay keeps
using the returned state/ref/callbacks so existing handlers and effects preserve
their behavior and dependency arrays.

Validated locally: git diff --check; focused ESLint --quiet on the touched files;
app tsc reports only the existing baseline errors, including the pre-existing
absoluteFillObject error in VideoPlayerOverlayImpl; launched the x86_64 debug app
on peartube-pixel through Metro, opened a playable feed video, and logcat showed
VideoPlayerOverlay/player media session buffering plus Player loaded with no fatal
JS/native crash. Screenshot showed the watch/player surface with P2P status,
actions, and comments visible.
